+++ OctopusParent/OrderServiceProvider/src/main/java/com/octopus/eureka/order/OrderFinancialController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -25,7 +26,7 @@ public class OrderFinancialController {
     OrderFinancialMapper orderFinancialMapper;
 
     @RequestMapping("/order/{orderSeq}")
-    public OrderFinancialDto getOrder(@PathVariable("orderSeq") double orderSeq) {
+    public OrderFinancialDto getOrder(@PathVariable("orderSeq") BigInteger orderSeq) {
         logger.info("请求参数orderSeq："+ orderSeq);
         return orderFinancialMapper.selectById(orderSeq);
     }
@@ -46,7 +47,7 @@ public class OrderFinancialController {
     }
 
     @RequestMapping("/order/delete/{orderSeq}")
-    public int deleteOrder(@PathVariable("orderSeq") double orderSeq) {
+    public int deleteOrder(@PathVariable("orderSeq") BigInteger orderSeq) {
         logger.info("请求参数orderSeq："+ orderSeq);
         return orderFinancialMapper.delete(orderSeq);
     }
@@ -55,7 +56,7 @@ public class OrderFinancialController {
     public OrderFinancialDto getAddOrder(@RequestBody  OrderFinancialDto orderFinancialDto) {
         int insert = orderFinancialMapper.insert(orderFinancialDto);
         if(insert > 0){
-            double orderSeq = orderFinancialDto.getOrderSeq();
+            BigInteger orderSeq = orderFinancialDto.getOrderSeq();
             logger.info("插入订单的orderSeq："+ orderSeq);
             return orderFinancialMapper.selectById(orderSeq);
         }
